requirements.txt
python-dotenv>=1.0.1
langchain>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.2.0
grandalf>=0.8
sendgrid>=6.11.0
streamlit>=1.38.0
serpapi>=0.1.5
langchain-google-genai>=2.0.0
pytest>=7.4.0
